+++ alicloud/resource_alicloud_fc_function.go
@@ -1,6 +1,7 @@
 package alicloud
 
 import (
+	"encoding/json"
 	"fmt"
 	"time"
 
@@ -74,6 +75,11 @@ func resourceAlicloudFCFunction() *schema.Resource {
 				Optional: true,
 			},
 
+			"environment_variables": {
+				Type:     schema.TypeMap,
+				Optional: true,
+			},
+
 			"handler": {
 				Type:     schema.TypeString,
 				Required: true,
@@ -125,9 +131,19 @@ func resourceAlicloudFCFunctionCreate(d *schema.ResourceData, meta interface{}) 
 		Timeout:      Int32Pointer(int32(d.Get("timeout").(int))),
 		MemorySize:   Int32Pointer(int32(d.Get("memory_size").(int))),
 	}
+	if variables := d.Get("environment_variables").(map[string]interface{}); len(variables) > 0 {
+		byteVar, err := json.Marshal(variables)
+		if err != nil {
+			return WrapError(err)
+		}
+		err = json.Unmarshal(byteVar, &object.EnvironmentVariables)
+		if err != nil {
+			return WrapError(fmt.Errorf("EnvironmentVariables must be type of map[string]string, err is %s", err.Error()))
+		}
+	}
 	code, err := getFunctionCode(d, client)
 	if err != nil {
-		return err
+		return WrapError(err)
 	}
 	object.Code = code
 	input.FunctionCreateObject = object
@@ -139,19 +155,19 @@ func resourceAlicloudFCFunctionCreate(d *schema.ResourceData, meta interface{}) 
 		})
 		if err != nil {
 			if IsExceptedErrors(err, []string{AccessDenied}) {
-				return resource.RetryableError(fmt.Errorf("Error creating function compute service got an error: %#v", err))
+				return resource.RetryableError(WrapError(err))
 			}
-			return resource.NonRetryableError(fmt.Errorf("Error creating function compute service got an error: %#v", err))
+			return resource.NonRetryableError(WrapError(err))
 		}
 		function, _ = raw.(*fc.CreateFunctionOutput)
 		return nil
 
 	}); err != nil {
-		return err
+		return WrapErrorf(err, DefaultErrorMsg, "fc_function", "CreateFunction", AliyunLogGoSdkERROR)
 	}
 
 	if function == nil {
-		return fmt.Errorf("Creating function compute function got a empty response: %#v.", function)
+		return WrapError(Error("Creating function compute function got a empty response"))
 	}
 
 	d.SetId(fmt.Sprintf("%s%s%s", serviceName, COLON_SEPARATED, *function.FunctionName))
@@ -165,7 +181,7 @@ func resourceAlicloudFCFunctionRead(d *schema.ResourceData, meta interface{}) er
 
 	split := strings.Split(d.Id(), COLON_SEPARATED)
 	if len(split) < 2 {
-		return fmt.Errorf("Invalid resource ID %s. Please check it and try again.", d.Id())
+		return WrapError(fmt.Errorf("Invalid resource ID %s. Please check it and try again.", d.Id()))
 	}
 
 	function, err := fcService.DescribeFcFunction(split[0], split[1])
@@ -174,7 +190,7 @@ func resourceAlicloudFCFunctionRead(d *schema.ResourceData, meta interface{}) er
 			d.SetId("")
 			return nil
 		}
-		return fmt.Errorf("DescribeFCFunction %s got an error: %#v", d.Id(), err)
+		return WrapError(err)
 	}
 
 	d.Set("service", split[0])
@@ -185,6 +201,7 @@ func resourceAlicloudFCFunctionRead(d *schema.ResourceData, meta interface{}) er
 	d.Set("runtime", function.Runtime)
 	d.Set("timeout", function.Timeout)
 	d.Set("last_modified", function.LastModifiedTime)
+	d.Set("environment_variables", function.EnvironmentVariables)
 
 	return nil
 }
@@ -192,7 +209,6 @@ func resourceAlicloudFCFunctionRead(d *schema.ResourceData, meta interface{}) er
 func resourceAlicloudFCFunctionUpdate(d *schema.ResourceData, meta interface{}) error {
 	client := meta.(*connectivity.AliyunClient)
 
-	d.Partial(true)
 	updateInput := &fc.UpdateFunctionInput{}
 	update := false
 
@@ -222,6 +238,17 @@ func resourceAlicloudFCFunctionUpdate(d *schema.ResourceData, meta interface{}) 
 		updateInput.Runtime = StringPointer(d.Get("runtime").(string))
 		d.SetPartial("runtime")
 	}
+	if d.HasChange("environment_variables") {
+		byteVar, err := json.Marshal(d.Get("environment_variables").(map[string]interface{}))
+		if err != nil {
+			return WrapError(err)
+		}
+		err = json.Unmarshal(byteVar, &updateInput.EnvironmentVariables)
+		if err != nil {
+			return WrapError(fmt.Errorf("EnvironmentVariables must be type of map[string]string, err is %s", err.Error()))
+		}
+		d.SetPartial("environment_variables")
+	}
 
 	if updateInput != nil || update {
 		split := strings.Split(d.Id(), COLON_SEPARATED)
@@ -229,7 +256,7 @@ func resourceAlicloudFCFunctionUpdate(d *schema.ResourceData, meta interface{}) 
 		updateInput.FunctionName = StringPointer(split[1])
 		code, err := getFunctionCode(d, client)
 		if err != nil {
-			return err
+			return WrapError(err)
 		}
 		updateInput.Code = code
 
@@ -237,11 +264,10 @@ func resourceAlicloudFCFunctionUpdate(d *schema.ResourceData, meta interface{}) 
 			return fcClient.UpdateFunction(updateInput)
 		})
 		if err != nil {
-			return fmt.Errorf("UpdateFunction %s got an error: %#v.", d.Id(), err)
+			return WrapErrorf(err, DefaultErrorMsg, d.Id(), "UpdateFunction", AliyunLogGoSdkERROR)
 		}
 	}
 
-	d.Partial(false)
 	return resourceAlicloudFCFunctionRead(d, meta)
 }
 
@@ -261,14 +287,14 @@ func resourceAlicloudFCFunctionDelete(d *schema.ResourceData, meta interface{}) 
 			if IsExceptedErrors(err, []string{ServiceNotFound, FunctionNotFound}) {
 				return nil
 			}
-			return resource.NonRetryableError(fmt.Errorf("Deleting function got an error: %#v.", err))
+			return resource.NonRetryableError(WrapErrorf(err, DefaultErrorMsg, d.Id(), "DeleteFunction", AliyunLogGoSdkERROR))
 		}
 
 		if _, err := fcService.DescribeFcFunction(split[0], split[1]); err != nil {
 			if NotFoundError(err) {
 				return nil
 			}
-			return resource.RetryableError(fmt.Errorf("While deleting function, getting function %s got an error: %#v.", d.Id(), err))
+			return resource.RetryableError(WrapErrorf(err, DefaultErrorMsg, d.Id(), "DeleteFunction", AliyunLogGoSdkERROR))
 		}
 		return nil
 	})
@@ -280,14 +306,14 @@ func getFunctionCode(d *schema.ResourceData, client *connectivity.AliyunClient) 
 	if filename, ok := d.GetOk("filename"); ok && filename.(string) != "" {
 		file, err := loadFileContent(filename.(string))
 		if err != nil {
-			return code, fmt.Errorf("Unable to load %q: %s", filename.(string), err)
+			return code, WrapError(fmt.Errorf("Unable to load %q: %s", filename.(string), err))
 		}
 		code.WithZipFile(file)
 	} else {
 		bucket, bucketOk := d.GetOk("oss_bucket")
 		key, keyOk := d.GetOk("oss_key")
 		if !bucketOk || !keyOk {
-			return code, fmt.Errorf("'oss_bucket' and 'oss_key' must all be set while using OSS code source.")
+			return code, WrapError(Error("'oss_bucket' and 'oss_key' must all be set while using OSS code source."))
 		}
 		code.WithOSSBucketName(bucket.(string)).WithOSSObjectName(key.(string))
 	}